pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.76.0",
    "python-dotenv>=1.0.1",
    "rich>=14.0.0",
]
